Crate reminisce

Reminisce is a lightweight library intended to be used for detecting and reading from joysticks at a low level across many different operating systems.

Scanning for joysticks

To scan for joysticks, a Backend must be created.

Scanning for events

To scan events, a Backend must be created and polled.

use reminisce::{Backend, Native};
let mut backend = Native::new();
for event in backend.iter() {
    println!("{:?}", event);
}

Reexports

pub use linux as native;

pub use native::Native;

Modules

linux

Structs

Poller
An iterator over a Backend's event queue.

StatefulJoystick
A joystick that tracks its state.

Enums

Event
An event emitted by a joystick

HatPos
A hat position on a joystick.

Traits

Backend
A lightweight Backend that tracks and polls all the available joysticks.

Joystick
A joystick or gamepad.

Type Definitions

Axis
A direction on a joystick.

Button
A button on a joystick.

Hat
A hat on a joystick.

JoystickIndex
A joystick index.