Enum reminisce::Event

pub enum Event {
    Connected(JoystickIndex),
    Disconnected(JoystickIndex),
    ButtonPressed(JoystickIndex, Button),
    ButtonReleased(JoystickIndex, Button),
    AxisMoved(JoystickIndex, Axis, f32),
    HatMoved(JoystickIndex, Hat, HatPos),
}

An event emitted by a joystick

Variants

Connected

Fired when a joystick is connected with its index.

Disconnected

Fired when a joystick is disconnected with its index.

ButtonPressed

Fired when a button is pressed with the joystick index and the button's index.

ButtonReleased

Fired when a button is released with the joystick index and the button's index.

AxisMoved

Fired when a axis is moved with the joystick index, axis index and its value, which is between -1 and 1

HatMoved

Fired when a hat is moved with the joystick index and the hat's index and position.