Trait reminisce::Backend

pub trait Backend: Sized {
    type Joystick: Joystick;
    fn new() -> Self;
    fn joysticks(&self) -> &[Self::Joystick];
    fn poll(&mut self) -> Option<Event>;

    fn num_joysticks(&self) -> usize { ... }
    fn iter(&mut self) -> Poller<Self> { ... }
}

A lightweight Backend that tracks and polls all the available joysticks.

Each Backend has its own event queue tied to the events of the joysticks.

use reminisce::*;
let mut backend = Native::new();
let mut dir = 0;
let mut left = 30;
for event in &mut backend {
    left -= 1;
    if(left <= 0) { break }
    match event {
        Event::AxisMoved(_, 0, nx) =>
            dir = -1,
        Event::AxisMoved(_, 1, ny) =>
            dir = 1,
        _ => ()
    }
}

Associated Types

type Joystick: Joystick

The kind of joystick this Backend tracks.

Required Methods

fn new() -> Self

Create a new Backend and scan for joysticks.

fn joysticks(&self) -> &[Self::Joystick]

Return a reference to the joysticks connected.

fn poll(&mut self) -> Option<Event>

Poll this Backend non-blockingly for events from any joysticks.

Provided Methods

fn num_joysticks(&self) -> usize

Returns the number of joysticks connected.

fn iter(&mut self) -> Poller<Self>

Iterate through the events that haven't been processed yet