Trait reminisce::Joystick

pub trait Joystick: Sized {
    type OpenError: Debug;
    fn open(index: JoystickIndex) -> Result<Self, Self::OpenError>;
    fn connected(&self) -> bool;
    fn id(&self) -> Cow<str>;
    fn index(&self) -> JoystickIndex;
    fn num_axes(&self) -> Axis;
    fn num_buttons(&self) -> Button;
    fn num_hats(&self) -> Hat;
    fn battery(&self) -> Option<f32>;
}

A joystick or gamepad.

Associated Types

type OpenError: Debug

The error that could be thrown while trying to open the joystick

Required Methods

fn open(index: JoystickIndex) -> Result<Self, Self::OpenError>

Attempts to open a joystick from its index

If an error occurs, this will return the textual representation of that error.

use reminisce::{Backend, Joystick, Native};
if let Ok(joystick) = <Native as Backend>::Joystick::open(0) {
    println!("{}", joystick.id())
} else {
    println!("No joystick plugged in")
}

fn connected(&self) -> bool

Check if the joystick is still connected.

fn id(&self) -> Cow<str>

Get the identifier of this joystick.

This is copy-on-write because it can be a borrowed or owned String, depending on the implementation.

fn index(&self) -> JoystickIndex

Get the index of this joystick.

fn num_axes(&self) -> Axis

Get the number of axes this joystick has

This is capped at 6 axes for now.

fn num_buttons(&self) -> Button

Get the number of buttons this joystick has

This is capped at 16 buttons currently.

fn num_hats(&self) -> Hat

Get the number of hats this joystick has

fn battery(&self) -> Option<f32>

Get the battery level of this joystick

Returns none if the joystick is wired or this operation is not supported by the backend