Trait reminisce::StatefulJoystick

pub trait StatefulJoystick: Joystick {
    fn get_axis(&self, index: u8) -> Option<i16>;
    fn get_button(&self, index: u8) -> Option<bool>;
    fn update(&mut self);

    fn get_normalised_axis(&self, index: u8) -> Option<f32> { ... }
}

A single joystick with its state saved

Required Methods

fn get_axis(&self, index: u8) -> Option<i16>

Get the value of a specific axis from its index

Typically the first two of these are the primary analog stick's x and y co-ordinates

fn get_button(&self, index: u8) -> Option<bool>

Get the value (if it is pressed or not) of a specific button

The first two buttons are usually the accept and back buttons

fn update(&mut self)

Update the state of this joystick

Provided Methods

fn get_normalised_axis(&self, index: u8) -> Option<f32>

Get the value of a specific axis normalised to between -1.0 and 1.0